myapp/requirements.txt
Flask==2.0.2
boto3==1.20.24
botocore==1.23.24
google-cloud-storage==1.43.0
azure-storage-blob==12.9.0
pyyaml
pytest
pytest-flask